made the sign out button into a component
