+++ src/components/SignOutButton.tsx
@@ -1,0 +1,34 @@
+// REMOVE SIGN OUT PAGE LATER
+
+import { createClient } from "@supabase/supabase-js";
+
+const supa = createClient(
+  import.meta.env.VITE_SUPABASE_URL!,
+  import.meta.env.VITE_SUPABASE_PUBLISHABLE_KEY!
+);
+
+function SignOutButton() {
+  const signOut = async () => {
+    await supa.auth.signOut();
+    window.location.href = "/"; // redirect to login
+  };
+
+  return (
+    <button
+        onClick={signOut}
+        style={{
+          backgroundColor: "#333",
+          color: "white",
+          border: "none",
+          borderRadius: "8px",
+          padding: "10px 20px",
+          cursor: "pointer",
+          marginBottom: "20px",
+        }}
+      >
+        Sign Out
+      </button>
+  );
+}
+
+export default SignOutButton;
